Added player fire cooldown, added some color to enemy/player lasers, fixed bug with player firing(more in the description), created prefab for player ship with fixes and added to all 3 levels.
The player laser bug was that since it was instantiating 2 objects sometimes 1 of the two lasers would not collide with the meteors or enemies and would hit another one behind it (this is not how it is intended). I created one spawn point with one laser fired instead in order to make it consistent with intended gameplay and created a prefab of the player ship and then updated all scenes with the correct ship.